Replace anonymous test class with stub resources.
This commit replaces inline anonymous classes for mocking `ResourceInterface` with pre-defined stub classes `FakeStubResource` and `FakeStub2Resource`. This change improves code readability and reusability in the test cases.
